Adding a custom operator multiplication method
This method is used to multiply operators on the same qubit subspace.

The existing combine_ops_to_mat method seems to expand the result matrix when there are multiple qubits.

Also, change a unit test which was set up incorrectly. Better to use known duration square pulses for testing to catch solver errors.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>