Added support for loading pulse library *only* json
As we have learnt that there could be a shortened form of the backend JSON which essentially only contains the pulse library and cmd-def, we need to properly handle that.

The new Json object is previously a child object hence just need to refactor the code and add an if-switch to handle it. Also, add a simple unit test to validate the parsing procedure.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>