Implemented a procedure to calibrate CNOT gate using CR pulses
The procedure is as follows:
 - With a system Hamiltonian defined (e.g. 2 transmon qubits), using a Gaussian square pulse on u channel (cross-resonance drive) to drive the target qubit to a X(pi/2) equivalent.
To achieve this, using nl-opt to optimize the pulse amplitude (see CR_Pulse_Calibration Ipython notebook)

- Now we have the CR pulse, but this pulse creates single-qubit rotation errors on both control and target qubits. Hence, we need to use generic U3 rotation gates to cancel those unwanted errors + add the remaining rotations (perfect CR = ZX^(1/2))
This step is numerically intensive since we need to use QPT (lots of simulations) for cost function and the parameter space is quite large (12 params).
See calibrate_backend.cpp for impl.

Note: for QuaC simulator, we only need to do this once for each Hamiltonian, then we have an optimal CNOT pulse implementation. The U3 gates will also be converted to pulses (single qubit gates are much easier to calibrate even with QPT)
Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>